+++ src/modules/mo-core/data-transfer-objects/cicdpipeline-log-entry.dto.ts
@@ -1,0 +1,10 @@
+import { Expose } from 'class-transformer';
+import { IsOptional } from 'class-validator';
+
+export class CiCdPipelineLogEntry {
+  @Expose()
+  id: number;
+
+  @Expose()
+  url: string;
+}