+++ src/modules/mo-namespace/data-transfer-objects/namespace-service/namespace-service-patch-request.dto.ts
@@ -1,6 +1,8 @@
-import { Expose } from 'class-transformer';
+import { Expose, Transform, Type } from 'class-transformer';
 import { NamespaceServiceCreateRequestDto } from './namespace-service-create-request.dto';
-import { IsNotEmpty, IsString, IsUUID } from 'class-validator';
+import { isArray, IsNotEmpty, IsOptional, IsString, IsUUID, ValidateNested } from 'class-validator';
+import { NamespaceServiceKubernetesSettingsPatchRequestDto } from './namespace-service-kubernetes-settings-patch-request.dto';
+import { NamespaceServiceEnvVarPatchRequestDto } from './namespace-service-envvar-patch-request.dto';
 
 export class NamespaceServicePatchRequestDto extends NamespaceServiceCreateRequestDto {
   @IsNotEmpty()
@@ -8,4 +10,17 @@ export class NamespaceServicePatchRequestDto extends NamespaceServiceCreateReque
   @IsUUID()
   @Expose()
   id: string;
+
+  @IsNotEmpty()
+  @Type(() => NamespaceServiceKubernetesSettingsPatchRequestDto)
+  @ValidateNested()
+  @Expose()
+  kubernetesSettings: NamespaceServiceKubernetesSettingsPatchRequestDto;
+
+  @IsOptional()
+  @Type(() => NamespaceServiceEnvVarPatchRequestDto)
+  @Transform(({ value }) => (value && isArray(value) ? value : []))
+  @ValidateNested()
+  @Expose()
+  envVars: NamespaceServiceEnvVarPatchRequestDto[];
 }