+++ src/modules/mo-notification/data-transfer-objects/k8s-notification.dto.ts
@@ -43,4 +43,7 @@ export class K8sNotificationDto {
   
   @Expose()
   positionOf: number;
+
+  @Expose()
+  buildId?: string;
 }